Add non-fhs-userenv devshell for Roc
Looks like to get a clean setup, I first need to run a `roc run` command
outside of the fhs-userenv. That will download the basic-cli program and
error out at some point. _Then_ I can run roc again inside of the
fhs-userenv and it will work. But only if we performed that failing step
first.